fix: 21 comments of ledger MR
