--- .changeset/red-bags-own.md
@@ -1,76 +1,0 @@
----
-"@ledgerhq/hw-transport-node-hid-singleton": minor
-"@ledgerhq/hw-transport-node-speculos-http": minor
-"dummy-wallet-app": minor
-"@ledgerhq/hw-transport-node-hid-noevents": minor
-"@ledgerhq/react-native-hw-transport-ble": minor
-"@ledgerhq/swift-bridge-hw-transport-ble": minor
-"dummy-live-app": minor
-"@ledgerhq/hw-transport-node-speculos": minor
-"@ledgerhq/swift-bridge-hw-app-solana": minor
-"@ledgerhq/swift-bridge-hw-app-eth": minor
-"@ledgerhq/hw-transport-node-ble": minor
-"@ledgerhq/hw-transport-node-hid": minor
-"@ledgerhq/hw-transport-web-ble": minor
-"@ledgerhq/hw-transport-mocker": minor
-"@ledgerhq/hw-transport-webhid": minor
-"@ledgerhq/hw-transport-webusb": minor
-"@ledgerhq/types-cryptoassets": minor
-"@actions/prepare-comment-screenshots": minor
-"@ledgerhq/hw-transport-http": minor
-"@ledgerhq/react-native-hid": minor
-"@actions/generate-bot-super-report": minor
-"@actions/get-workflow-artifacts-id": minor
-"@ledgerhq/hw-app-algorand": minor
-"@ledgerhq/hw-app-exchange": minor
-"@ledgerhq/hw-app-polkadot": minor
-"@actions/generate-release-message": minor
-"@ledgerhq/hw-app-cosmos": minor
-"@ledgerhq/hw-app-helium": minor
-"@ledgerhq/hw-app-solana": minor
-"@ledgerhq/types-devices": minor
-"@ledgerhq/cryptoassets": minor
-"@ledgerhq/hw-app-tezos": minor
-"@ledgerhq/hw-transport": minor
-"@ledgerhq/hw-app-near": minor
-"@actions/live-common-affected": minor
-"@ledgerhq/hw-app-btc": minor
-"@ledgerhq/hw-app-eth": minor
-"@ledgerhq/hw-app-str": minor
-"@ledgerhq/hw-app-trx": minor
-"@ledgerhq/hw-app-vet": minor
-"@ledgerhq/hw-app-xrp": minor
-"@ledgerhq/types-live": minor
-"@actions/change-product-name": minor
-"@actions/turborepo-gh-cache": minor
-"@actions/get-package-infos": minor
-"@actions/submit-bot-report": minor
-"@ledgerhq/devices": minor
-"@ledgerhq/errors": minor
-"@ledgerhq/crypto-icons-ui": minor
-"@actions/turbo-affected": minor
-"@ledgerhq/logs": minor
-"@actions/upload-images": minor
-"ledger-live-desktop": minor
-"metro-extra-config": minor
-"live-mobile": minor
-"@ledgerhq/live-common": minor
-"@ledgerhq/native-ui": minor
-"@ledgerhq/ui-shared": minor
-"@ledgerhq/icons-ui": minor
-"@ledgerhq/react-ui": minor
-"@ledgerhq/coin-framework": minor
-"@ledgerhq/domain-service": minor
-"@ledgerhq/live-portfolio": minor
-"esbuild-utils": minor
-"@ledgerhq/coin-polkadot": minor
-"live-common-tools": minor
-"live-github-bot": minor
-"ledger-libs": minor
-"actions-root": minor
-"@ledgerhq/live-promise": minor
-"@ledgerhq/live-cli": minor
-"@ledgerhq/live-env": minor
----
-
-vechain integration
