+++ libs/ledger-live-common/src/families/vechain/js-getTransactionStatus.ts
@@ -9,10 +9,10 @@ import {
 import type { TransactionStatus } from "./types";
 import type { Transaction } from "./types";
 import { Account } from "@ledgerhq/types-live";
-import { calculateTotalSpent, calculateTransactionInfo } from "./utils/calculateTransactionInfo";
 import { isValid } from "./utils/address-utils";
 import BigNumber from "bignumber.js";
 import { NotEnoughVTHO } from "./errors";
+import { calculateTotalSpent, calculateTransactionInfo } from "./utils/transaction-utils";
 
 const getTransactionStatus = async (
   account: Account,
